requirements.txt
Django==3.0.2
django-filter==2.2.0
djangorestframework==3.11.0
Markdown==3.2.1
psycopg2-binary==2.8.4
persian==0.4.0